Add ApiExceptionHandler for uncaught Refit exceptions AB#16709 (#6120)
