+++ Apps/Common/src/AspNetConfiguration/Modules/ExceptionHandling.cs
@@ -16,7 +16,7 @@
 namespace HealthGateway.Common.AspNetConfiguration.Modules
 {
     using System.Diagnostics.CodeAnalysis;
-    using HealthGateway.Common.ErrorHandling;
+    using HealthGateway.Common.ErrorHandling.ExceptionHandlers;
     using Microsoft.AspNetCore.Builder;
     using Microsoft.Extensions.DependencyInjection;
 
@@ -32,6 +32,7 @@ namespace HealthGateway.Common.AspNetConfiguration.Modules
         /// <param name="services">The service collection to add forward proxies into.</param>
         public static void ConfigureProblemDetails(IServiceCollection services)
         {
+            services.AddExceptionHandler<ApiExceptionHandler>();
             services.AddExceptionHandler<ValidationExceptionHandler>();
             services.AddExceptionHandler<DefaultExceptionHandler>();
             services.AddProblemDetails();
